// english-openai/apps/server/src/infractructure/db/enums/db-models.enum.ts
export enum EDbModels {
  User = "User",
  Sentence = "Sentence",
  SentenceCheck = "SentenceCheck",
  Sharing = "Sharing",
  TrainingTask = "TrainingTask",
  TrainingSession = "TrainingSession",
  Translation = "Translation",
  Token = "Token",
  Word = "Word",
  WordSet = "WordSet",
  UserSettings = "UserSettings",
}